+++ ebookManager/src/com/thomsonreuters/uscl/ereader/mgr/web/controller/group/edit/Title.java
@@ -3,11 +3,13 @@ package com.thomsonreuters.uscl.ereader.mgr.web.controller.group.edit;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class Title {
     private String titleId;
     private String proviewName;
-    private Integer version;
+    private BigInteger version;
     private int numberOfParts;
 }
